Move user creation methods to privaterelay
