+++ emails/tests/mgmt_send_welcome_emails_tests.py
@@ -12,8 +12,8 @@ from allauth.socialaccount.models import SocialAccount
 from botocore.exceptions import ClientError
 
 from emails.models import Profile
-from emails.tests.models_tests import make_free_test_user
 from privaterelay.ftl_bundles import main as ftl_bundle
+from privaterelay.tests.utils import make_free_test_user
 
 COMMAND_NAME = "send_welcome_emails"
